# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/tasks:
    get:
      tags:
        - Tasks
      summary: List Tasks
      description: Get a paginated list of Tasks with optional filtering and sorting
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tasks
      summary: Create Task
      description: Create a new Task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/timesheet_details:
    get:
      tags:
        - TimesheetDetails
      summary: List TimesheetDetails
      description: Get a paginated list of TimesheetDetails with optional filtering and sorting
      operationId: listTimesheetDetails
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetDetailList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimesheetDetails
      summary: Create TimesheetDetail
      description: Create a new TimesheetDetail
      operationId: createTimesheetDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimesheetDetailRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Task:
      type: object
      required:
        - id
        - company_id
        - project_id
        - subject
        - status
        - expected_time
        - progress
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        parent_task_id:
          nullable: true
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        task_type:
          nullable: true
          type: string
          maxLength: 80
        status:
          type: string
        expected_time:
          type: string
        progress:
          type: string
        metadata:
          type: object
...
    CreateTaskRequest:
      type: object
      required:
        - company_id
        - project_id
        - subject
        - status
        - expected_time
        - progress
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        parent_task_id:
          nullable: true
          type: string
          format: uuid
        subject:
          type: string
          maxLength: 200
        task_type:
          nullable: true
          type: string
          maxLength: 80
        status:
          type: string
        expected_time:
          type: string
        progress:
          type: string
        metadata:
          type: object
...
    TaskList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    TimesheetDetail:
      type: object
      required:
        - id
        - company_id
        - timesheet_id
        - hours
        - billing_rate
        - costing_rate
        - is_billable
        - billable_amount
        - costing_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        timesheet_id:
          type: string
          format: uuid
        activity_type_id:
          nullable: true
          type: string
          format: uuid
        task_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 280
        hours:
          type: string
        billing_rate:
          type: string
        costing_rate:
          type: string
        is_billable:
          type: boolean
        billable_amount:
          type: string
        costing_amount:
          type: string
        metadata:
          type: object
...
    CreateTimesheetDetailRequest:
      type: object
      required:
        - company_id
        - timesheet_id
        - hours
        - billing_rate
        - costing_rate
        - is_billable
        - billable_amount
        - costing_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        timesheet_id:
          type: string
          format: uuid
        activity_type_id:
          nullable: true
          type: string
          format: uuid
        task_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 280
        hours:
          type: string
        billing_rate:
          type: string
        costing_rate:
          type: string
        is_billable:
          type: boolean
        billable_amount:
          type: string
        costing_amount:
          type: string
        metadata:
          type: object
...
    TimesheetDetailList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimesheetDetail'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []