# src/presentation/http/timesheet_detail_handler.rs
//! TimesheetDetail REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{TimesheetDetailService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateTimesheetDetailDto, UpdateTimesheetDetailDto, PatchTimesheetDetailDto, TimesheetDetailResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum TimesheetDetailError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for TimesheetDetailError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for TimesheetDetailError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "TIMESHEETDETAIL_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "TIMESHEETDETAIL_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TIMESHEETDETAIL_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TIMESHEETDETAIL_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for TimesheetDetail.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /timesheet_details | List with pagination |
/// | POST | /timesheet_details | Create new |
/// | GET | /timesheet_details/:id | Get by ID |
/// | PUT | /timesheet_details/:id | Full update |
/// | PATCH | /timesheet_details/:id | Partial update |
/// | DELETE | /timesheet_details/:id | Soft delete |
/// | POST | /timesheet_details/bulk | Bulk create |
/// | POST | /timesheet_details/upsert | Upsert |
/// | GET | /timesheet_details/trash | List deleted |
/// | POST | /timesheet_details/:id/restore | Restore |
/// | DELETE | /timesheet_details/empty | Empty trash |
/// | GET | /timesheet_details/:id/deleted | Get deleted by ID |
/// | DELETE | /timesheet_details/trash/:id | Permanent delete from trash |
/// | GET | /timesheet_details/count | Count active entities |
/// | GET | /timesheet_details/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(TimesheetDetailService::with_repository(repository));
/// let router = create_timesheet_detail_routes(service);
/// ```
pub fn create_timesheet_detail_routes(service: Arc<TimesheetDetailService>) -> Router {
    BackboneCrudHandler::<TimesheetDetailService, TimesheetDetail, CreateTimesheetDetailDto, UpdateTimesheetDetailDto, TimesheetDetailResponseDto>::routes(
        service,
        "/timesheet_details",
    )
}

/// Create Axum router with only the read (GET) endpoints for TimesheetDetail.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_timesheet_detail_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_timesheet_detail_read_routes(service: Arc<TimesheetDetailService>) -> Router {
    BackboneCrudHandler::<TimesheetDetailService, TimesheetDetail, CreateTimesheetDetailDto, UpdateTimesheetDetailDto, TimesheetDetailResponseDto>::read_routes(
        service,
        "/timesheet_details",
    )
}

/// Create Axum router with only the write (mutation) endpoints for TimesheetDetail.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_timesheet_detail_write_routes(service: Arc<TimesheetDetailService>) -> Router {
    BackboneCrudHandler::<TimesheetDetailService, TimesheetDetail, CreateTimesheetDetailDto, UpdateTimesheetDetailDto, TimesheetDetailResponseDto>::write_routes(
        service,
        "/timesheet_details",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_timesheet_detail_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<TimesheetDetailService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_timesheet_detail_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}